All things working
All of the functionalities asked are working correctly. 
The user interface is in spanish and the user can create different billboards that are going to be saved, atomatically, in a .temp file. Also, the user can import Datos1 and Datos2 and visualize all the billboards created. At the end, the user can generate the dangerous report that is saved in a .txt file.